+++ .vscode/settings.json
@@ -1,0 +1,3 @@
+{
+    "python.pythonPath": "C:\\Users\\Reha_2\\AppData\\Local\\Microsoft\\WindowsApps\\python.exe"
+}
