+++ devconnect/auth/__init__.py
@@ -1,5 +1,5 @@
 from flask import Blueprint
 
-bp = Blueprint('auth', __name__)
+bp = Blueprint('auth', __name__, template_folder="templates", static_folder="static", static_url_path="/auth/static")
 
 from devconnect.auth import routes
