+++ devconnect/auth/templates/login.html
@@ -1,0 +1,19 @@
+{% extends "base.html" %}
+{% import "bootstrap/wtf.html" as wtf %}
+
+{% block title %}
+Login
+{% endblock %}
+
+{% block page_content %}
+<h3 class="reha">log in</h3>
+<form method="POST" action="/login">
+    {{ form.hidden_tag() }}
+    {{ wtf.form_field(form.username) }}
+    {{ wtf.form_field(form.password) }}
+    {{ wtf.form_field(form.remember) }}
+    <button class="btn btn-primary" type="submit">Log In</button>
+</form>
+<br>
+<p><a href="/signup">Don't have an account? Sign up!</a></p>
+{% endblock %}
