+++ devconnect/auth/templates/signup.html
@@ -7,13 +7,12 @@ Sign Up
 
 {% block scripts %}
 {{ super() }}
-<script src="{{ url_for('static', filename='form.js') }}"></script>
+<script src="{{ url_for('auth.static', filename='form.js') }}"></script>
 {% endblock %}
 
 {% block styles %}
 {{ super() }}
-<link rel="stylesheet"
-      href="{{ url_for('static', filename='form.css') }}">
+<link rel="stylesheet" href="{{ url_for('auth.static', filename='form.css') }}">
 {% endblock %}
 
 {% block content %}
@@ -59,4 +58,3 @@ Sign Up
     </form>
 </div>
 {% endblock %}
-
