+++ devconnect/templates/base.html
@@ -11,17 +11,17 @@
         <div class="navbar-collapse collapse">
             <ul class="nav navbar-nav navbar-left">
                 <li class="nav-item">
-                    <a class="nav-link" href="{{url_for('home')}}">Home</a>
+                    <a class="nav-link" href="{{url_for('home.index')}}">Home</a>
                 </li>
             </ul>
             <ul class="nav navbar-nav navbar-right">
                 {% if current_user.is_authenticated %}
                 <li class="nav-item">
-                    <a class="nav-link" href="{{url_for('create_post')}}">Create</a>
+                    <a class="nav-link" href="{{url_for('create_post.create_post')}}">Create</a>
                 </li>
                 <li class="nav-item">
                     <a class="nav-link"
-                        href="{{ url_for('view_user', username=current_user.username) }}">{{ current_user.username }}</a>
+                        href="{{ url_for('profile.view_user', username=current_user.username) }}">{{ current_user.username }}</a>
                 </li>
                 <li class="nav-item">
                     <a class="nav-link" href="{{url_for('auth.logout')}}">Logout</a>
