--- devconnect/templates/index.html
@@ -1,18 +1,0 @@
-{% extends "base.html" %}
-{% import "bootstrap/wtf.html" as wtf %}
-
-{% block title %}
-Sign Up
-{% endblock %}
-
-{% block content %}
-<div class="container" style="width: 50%">
-    <h3 class="text-center">whatever dude</h3>
-    <div class="row">
-        <div class="text-center">
-            <a href="/signup" class="btn btn-primary">sign up</a>
-            <a href="/login" class="btn btn-primary">log in</a>
-        </div>
-    </div>
-</div>
-{% endblock %}
